+++ src/commands/handlers/world/index.ts
@@ -3,5 +3,6 @@ export { getCompendiumsHandler } from './GetCompendiumsHandler';
 export { getCompendiumHandler } from './GetCompendiumHandler';
 export { getCompendiumIndexHandler } from './GetCompendiumIndexHandler';
 export { searchCompendiumHandler } from './SearchCompendiumHandler';
+export { searchCompendiumsHandler } from './SearchCompendiumsHandler';
 export { getCompendiumDocumentHandler } from './GetCompendiumDocumentHandler';
 export { importFromCompendiumHandler } from './ImportFromCompendiumHandler';